+++ backend/db/models/review.js
@@ -45,7 +45,7 @@ module.exports = (sequelize, DataTypes) => {
       allowNull: false
     },
     stars: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.DECIMAL(5,2),
       allowNull: false,
       validate: {
         min: 1,
